reading and overwrite file works, resizing window works, framebuffer still not completed (problem with resizing)

+++ LillaSpelprojektet/global_settings.h
@@ -102,7 +102,7 @@ public:
 	}
 	void UpdateValuesFromFile();
 	float ValueOf(std::string setting_name);
-	void UpdateResolutionValues(float width, float height);
+	void UpdateResolutionValues(int width, int height);
 
 private:
 	GlobalSettings() {};
